Fixes iOS 6 Warnings for ServerlessDemo

+++ Xcode/ServerlessDemo/Classes/RootViewController.m
@@ -99,7 +99,7 @@ static const int ddLogLevel = LOG_LEVEL_VERBOSE;
 	if (numberOfUnreadMessages > 0)
 	{
 		UIButton *unreadIndicator = (UIButton *)cell.accessoryView;
-		NSString *unreadTitle = [NSString stringWithFormat:@"%lu", numberOfUnreadMessages];
+		NSString *unreadTitle = [NSString stringWithFormat:@"%lu", (unsigned long)numberOfUnreadMessages];
 		
 		[unreadIndicator setTitle:unreadTitle forState:UIControlStateDisabled];
